Move widgets to right place
